+++ src/app/pages/settings/settings.component.ts
@@ -314,7 +314,7 @@ export class SettingsPageComponent {
   }
 
   getAppVersion() {
-    return environment.version || '1.0.3';
+    return environment.version || '1.0.4';
   }
 
   getPlatform() {